Always use the OrderID from Shipstation on the notify hook
Instead of trying to find by number which could be either shipment
number or order number (and it's not clear as to what to do when it's an
order number that has multiple shipments). The xml builder for
shipstation already wires shipment id as order id, so we should continue
relying on the order id returned which is the shipment id. Allows us to
arbitrarily flip between displaying order number or shipment number in
Shipstation without affecting the functionality